+++ src/components/NewSplitInput.vue
@@ -1,7 +1,15 @@
 <template>
-  <div class="flex items-center [&>*]:mx-2">
+  <div
+    class="
+      grid
+      grid-cols-transactionListItem
+      items-center
+      [&>*]:mx-2
+    "
+  >
     <div />
-    <div class="flex-auto">
+    <div />
+    <div>
       <SeamlessInput
         :value="description"
         :placeholder="$t('general.description')"
@@ -14,7 +22,7 @@
         @account-changed="(accountId) => (destAccountId = accountId)"
       />
     </div>
-    <div class="flex-shrink-0">
+    <div>
       <CurrencyInput
         :value="value"
         :placeholder="$t('general.value')"